Set up a basic React application
This replaces the earlier basic Express application with a completely
custom setup that includes Webpack, Babel and React. Since I'm a newbie
this setup is still suspect and liable to change radically. The README
file has been updated to included references to articles that I
followed in setting this repository up.